replace EvDvrpFleetProvider with EvDvrpFleetModule
